Fix Pill import cycles
