+++ apps/web/src/components/views/elements/ReplyChain.tsx
@@ -19,7 +19,8 @@ import { getUserNameColorClass } from "../../../utils/FormattingUtils";
 import { Action } from "../../../dispatcher/actions";
 import Spinner from "./Spinner";
 import ReplyTile from "../rooms/ReplyTile";
-import { Pill, PillType } from "./Pill";
+import { Pill } from "./Pill";
+import { PillType } from "./PillType";
 import AccessibleButton from "./AccessibleButton";
 import { getParentEventId, shouldDisplayReply } from "../../../utils/Reply";
 import RoomContext from "../../../contexts/RoomContext";
