+++ apps/web/src/hooks/usePermalinkEvent.ts
@@ -9,7 +9,7 @@ Please see LICENSE files in the repository root for full details.
 import { MatrixEvent, type Room } from "matrix-js-sdk/src/matrix";
 import { useEffect, useState } from "react";
 
-import { PillType } from "../components/views/elements/Pill";
+import { PillType } from "../components/views/elements/PillType";
 import { MatrixClientPeg } from "../MatrixClientPeg";
 import { type PermalinkParts } from "../utils/permalinks/PermalinkConstructor";
