+++ apps/web/src/renderer/pill.tsx
@@ -13,7 +13,8 @@ import { textContent } from "domutils";
 import reactStringReplace from "react-string-replace";
 import { PushProcessor } from "matrix-js-sdk/src/pushprocessor";
 
-import { Pill, PillType } from "../components/views/elements/Pill";
+import { Pill } from "../components/views/elements/Pill";
+import { PillType } from "../components/views/elements/PillType";
 import { parsePermalink } from "../utils/permalinks/Permalinks";
 import { type PermalinkParts } from "../utils/permalinks/PermalinkConstructor";
 import { hasParentMatching, type RendererMap, type ParentNode } from "./utils.tsx";
